requirements.txt
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.3.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
requests==2.31.0